+++ Recipies/BackEnd/ClientApp/src/app/recipe-detailed-info/recipe-directions/recipe-directions.component.html
@@ -1,24 +1,6 @@
-<h2 style="margin-top:3rem;">Cooking process</h2>
-<ol>
-  <li>Ek.Bowl Me Cornflour Ley.Usey Acche Sey Ghol Ready Kar Ley.</li>
-  
-  <li>Without Lumps..Fir Ek Pan Ley.Usme Sugar Daal Dey ,1/2cup Water Daal Kr Bs Sugar Dissolved Hone TK.Pakai..Then Add Cornflour Liquid.
-    </li>
-  <li>Usey Bilkul Low Flame Pr Rakhe Aor Lagatar Chalte Rhe.Jb Vo Thoda Thick Ho Jaey..TB 2tbsp Ghee Add Kre Fir Chalte Rhe.
-    </li>
-  <li>Isi Tarah Aapko Fir 2baar 3_3 Tbsp Ghee Chalte Huey Add Karna H.
-    </li>
-  <li>Fir Colour Add Kre .1lemon Juice Add Kre..Fir Isey TB TK Chaley Jb TK Ye Transparent N Ho Jaey..
-    </li>
-  <li>Or Deakhne Me Lgne Lge Ki Ab Ye Jum Sakta H
-    </li>
-  <li>Fir Ek Plate Me Ghee Greece Kr Isey Acche Sey Spread Kar Dey.
-    </li>
-  <li>Fir 1 Hour K Ley Room Temperature Or Chod Dey
-    </li>
-  <li>Uske Baad Aap Apne Choice K Shape K According Cut Kre.
-    </li>
-    <li>Ho Gya Ready</li>
+<h2 id="cp" style="margin-top:3rem;">Cooking process:</h2>
+<ol *ngFor="let item of Directions">
+ <li>{{item}}</li>
 </ol>
