+++ Recipies/BackEnd/ClientApp/src/app/recipe-detailed-info/recipe-directions/recipe-directions.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit,Input} from '@angular/core';
 
 @Component({
   selector: 'app-recipe-directions',
@@ -7,9 +7,13 @@ import { Component, OnInit } from '@angular/core';
 })
 export class RecipeDirectionsComponent implements OnInit {
 
+
+  @Input() Directions: string[];
+
   constructor() { }
 
   ngOnInit() {
+    
   }
 
 }
