--- Recipies/BackEnd/ClientApp/src/app/recipe-detailed-info/recipe-galleria/recipe-galleria.component.css
@@ -1,1 +1,0 @@
-
