+++ Recipies/BackEnd/ClientApp/src/app/recipe-detailed-info/recipe-info/recipe-info.component.html
@@ -5,7 +5,7 @@
      <div class="col-xl-8 col-lg-8 col-md-12 col-sm-12"  style="border-right-style:solid;border-color: #A6ABB3;border-width: 0.1rem;;">
 
         <!--Topic -->
-       <h1 style="text-align: center">Chettinad Chicken</h1>
+       <h1 id="htop" style="text-align: center">Chettinad Chicken</h1>
 
         <!-- galleria -->
         <app-recipe-galleria ></app-recipe-galleria>
@@ -16,20 +16,6 @@
             <p class="p-description">Chicken Chettinad or Chettinad chicken is a classic Indian recipe, from the cuisine of Chettinad. It consists of chicken marinated in yogurt, turmeric and a paste of red chillies, kalpasi, coconut, poppy seeds, coriander seeds, cumin seeds, fennel seeds, black pepper, ground nuts, onions, garlic and gingelly oil. It is served hot and garnished with coriander leaves, accompanied with boiled rice or paratha.</p>
         </div>
         
-
-
-        <!-- Cooking Time-->
-        <!-- <h2 style="margin-top:2rem">Cooking Time:</h2>
-        
-        <div class="ui-g" >
-            
-            <div class="ui-g-12 ui-md-6 ui-lg-3">
-            <p class="pt">Preparation Time : 4</p>
-            </div>
-            <div class="ui-g-12 ui-md-6 ui-lg-3"><p class="ct">Cook Time : 20</p></div>
-            <div class="ui-g-12 ui-md-6 ui-lg-3"><p class="tt">Total Time : 24</p></div>
-      
-        </div> -->
 
 
         <!-- Ingradients -->
@@ -46,11 +32,12 @@
         <!-- <app-recipe-raiting></app-recipe-raiting> -->
 
 
+
         <!-- Date of publication -->
         <p class = "dp">Date of publication: 21.04.19</p>
 
+
         <!--Count views -->
-       
         <p class="cv">Count views: 15</p>
