+++ Recipies/BackEnd/ClientApp/src/app/recipe-detailed-info/recipe-info/recipe-info.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Recipe } from '../recipe';
-import { RecipeService } from '../recipe.service';
+import { RecipeService } from '../../services/recipe.service';
 import { Observable } from 'rxjs';
 
 @Component({
@@ -11,13 +11,14 @@ import { Observable } from 'rxjs';
 export class RecipeInfoComponent implements OnInit {
   
   
-  recipe:Observable<Recipe>;
+  public recipe:Observable<Recipe>;
   constructor() {
-    // private Rs:RecipeService,private RecipeId:string
+    // private Rs:RecipeService,private RecipeId
    }
   
   ngOnInit() {
    //this.recipe=this.Rs.getRecipeById(this.RecipeId);
+   
   }
   
   }
