+++ Recipies/BackEnd/ClientApp/src/app/recipe-detailed-info/recipe-ingradients/recipe-ingradients.component.html
@@ -1,13 +1,6 @@
-<h2>Ingradients</h2>
-<ul>
-  <li>Cornflour 1/2 Cups</li>
-  <li>Sugar 1 Cups</li>
-  <li>Cardomom Powder 1/2 Tsp</li>
-  <li>Some Cashew Nuts 2 Tbsp</li>
-  <li>Orange Food Colour 1 Pinch</li>
-  <li>Lemon Juice 2 Tsp</li>
-  <li>Ghee 1/2 Bowls </li>
-  <li>Water 1/2 Bowls</li>
+<h2 id="hing">Ingredients:</h2>
+<ul *ngFor="let ingredient of Ingredients">
+ <li>{{ingredient}}</li>
 </ul>
