+++ Recipies/BackEnd/ClientApp/src/app/recipe-detailed-info/recipe-ingradients/recipe-ingradients.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Input } from '@angular/core';
 
 @Component({
   selector: 'app-recipe-ingradients',
@@ -7,6 +7,7 @@ import { Component, OnInit } from '@angular/core';
 })
 export class RecipeIngradientsComponent implements OnInit {
 
+  @Input() Ingredients: string[];
   constructor() { }
 
   ngOnInit() {
