--- Recipies/BackEnd/ClientApp/src/app/recipe-detailed-info/recipe.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-
-import { RecipeService } from './recipe.service';
-
-describe('RecipeService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [RecipeService]
-    });
-  });
-
-  it('should be created', inject([RecipeService], (service: RecipeService) => {
-    expect(service).toBeTruthy();
-  }));
-});
